requirements.txt
numpy>=1.26
Pillow>=10.0

